Add test case for token generation Go fmt.

+++ config.go
@@ -3,20 +3,19 @@ package main
 import "github.com/satori/go.uuid"
 
 type Config struct {
-  port string
-  redisPort string
-  UUIDNamespace uuid.UUID
-  tokenKey string
-  tokenLifetime int
+	port          string
+	redisPort     string
+	UUIDNamespace uuid.UUID
+	tokenKey      string
+	tokenLifetime int64
 }
 
 var uuidName, _ = uuid.FromString("27d03927-7c8f-469e-8ba1-68a376d43cc9")
 
 var LocalConfig = Config{
-  port: ":5656",
-  redisPort: ":6379",
-  UUIDNamespace: uuidName,
-  tokenKey: "lshjdfsdjhf",
-  tokenLifetime: 3600,
+	port:          ":5656",
+	redisPort:     ":6379",
+	UUIDNamespace: uuidName,
+	tokenKey:      "lshjdfsdjhf",
+	tokenLifetime: 3600,
 }
-
